Fixed compiler warnings

+++ Synthie/CFlange.cpp
@@ -9,7 +9,7 @@ CFlange::CFlange(int channels, double sampleRate, double samplePeriod) : CEffect
 	m_amplitude = 1;
 
 	m_bufferIndex = 0;
-	m_bufferSize = std::ceil(m_channels * m_amplitude * m_sampleRate);
+	m_bufferSize = (int)std::ceil(m_channels * m_amplitude * m_sampleRate);
 
 	m_feedback = 0.5;
 	m_wetness = 0.5;
@@ -28,7 +28,7 @@ void CFlange::Process(const double* frameIn, double* frameOut, const double& tim
 	// distortion in samples.
 	const double waveform = m_amplitude * sin(m_phase * 2 * PI);
 
-	const int delayed = std::ceil(m_amplitude * m_sampleRate);
+	const int delayed = (int)std::ceil(m_amplitude * m_sampleRate);
 	m_phase += m_frequency * m_samplePeriod;
 
 	// Use frameHistory as a circular buffer.
@@ -37,7 +37,7 @@ void CFlange::Process(const double* frameIn, double* frameOut, const double& tim
 		const double input = frameIn[c];
 		double output = frameOut[c];
 
-		int i = std::ceil(std::fmod(m_bufferIndex - delayed, m_bufferSize));
+		int i = (int)std::ceil(std::fmod(m_bufferIndex - delayed, m_bufferSize));
 
 		// Avoid underflow.
 		if (i < 0)
@@ -53,7 +53,7 @@ void CFlange::Process(const double* frameIn, double* frameOut, const double& tim
 			frameIn[c] * (1 - m_feedback);
 
 		// Next index
-		m_bufferIndex = std::fmod(++m_bufferIndex, m_bufferSize);
+		m_bufferIndex = (int)std::fmod(++m_bufferIndex, m_bufferSize);
 
 		// Write output
 		frameOut[c] = output;
